chore: improve extension handling
Extensions will be loaded in parallel rather than one at a time, significantly improving how quickly results are fetched when the app first starts.
Implements better error handling visually showing when extensions fail to validate in the UI, allowing the user to manually revalidate them.
Implements visual indicators to show extensions that are currently in the process of loading.