feat: launch external player from native
Adds a button in the native player that launches an external player for video playback if a player path is defined on desktop or if you are on Android.
This button is usable with the "Enable External Player" setting disabled. This gives flexibility in using the native player when possible but can use the external player easily if an unsupported codec error occurs.